requirements.txt
git+https://github.com/codelif/myaakash
flask
flask-login
flask-sqlalchemy
python-dotenv
psycopg2-binary
